refactor(fn): changed import paths
